Refactor SentencePart to Clause
