+++ packages/yoastseo/spec/languageProcessing/languages/es/helpers/internal/getParticiplesSpec.js
@@ -1,9 +1,9 @@
 import getParticiples from "../../../../../../src/languageProcessing/languages/es/helpers/internal/getParticiples.js";
-import SentencePart from "../../../../../../src/values/SentencePart.js";
+import Clause from "../../../../../../src/values/Clause.js";
 
 describe( "Test for matching Spanish participles", function() {
 	it( "returns matched irregular participles.", function() {
-		const mockSentence = new SentencePart( "fueron agasajados con un lunch.", [ "fueron" ] );
+		const mockSentence = new Clause( "fueron agasajados con un lunch.", [ "fueron" ] );
 		const sentencePartText = mockSentence.getSentencePartText();
 		const auxiliaries = mockSentence.getAuxiliaries();
 		const foundParticiples = getParticiples( sentencePartText, auxiliaries );
@@ -16,7 +16,7 @@ describe( "Test for matching Spanish participles", function() {
 	} );
 
 	it( "returns an empty array when there is no participle or when the sentence is empty.", function() {
-		const mockSentence = new SentencePart( "Yo como una manzana.", [] );
+		const mockSentence = new Clause( "Yo como una manzana.", [] );
 		const sentencePartText = mockSentence.getSentencePartText();
 		const auxiliaries = mockSentence.getAuxiliaries();
 		expect( getParticiples( sentencePartText, auxiliaries ) ).toEqual( [] );
