+++ packages/yoastseo/src/languageProcessing/languages/hu/helpers/getSentenceParts.js
@@ -1,6 +1,6 @@
 import splitSentence from "../../../helpers/passiveVoice/periphrastic/getSentencePartsSplitOnStopwords";
 import arrayToRegex from "../../../helpers/regex/createRegexFromArray";
-import SentencePart from "../values/SentencePart";
+import SentencePart from "../values/Clause";
 import auxiliaries from "../config/internal/auxiliaries.js";
 import stopWords from "../config/stopWords.js";
